Added feature that allows us to play video from byte array. Cleaned up the API interface

+++ src/entities/Compass.java
@@ -1,5 +1,5 @@
 package entities;
 
 public interface Compass {
-	public float getAzimuth();
+	public double getAzimuth();
 }
